refactor: reorganize report generation helpers and improve import structure
